Remove all fields that aren't relevant to the discount mapping
The type, variableValues, hasNextPage, and hasPreviousPage fields are all
added on by the GraphQL JS client library, so we don't need to test those.

The __typename fields we aren't going to be changing, so we can ignore them as well
for now when rendering the outputs.

This way we can more directly compare the outputs, since it will only render the fields
we are actually going to be mapping