Finish getting payload mapping function working
Previously I had thought that order-level discounts were allocated based
on the cost for an individual *item* (within each line item). This was
incorrect and we don't need to know/take into account the quantity for each
line item for our calculations.

Testing this is a bit finicky because tests will fail even if it is "correct"
for our purposes. In order to get tests to pass, we will need to implement sorting
on the resulting arrays (of line items and discount allocations within those line items)
so that we can accurately programmatically compare the expected vs actual results.
This sorting logic has not yet been implemented.